renamed test namespaces
