+++ tests/fuzzing/Tests/ReadOnlySpan/ReadOnlySpanTests.Split.cs
@@ -1,7 +1,7 @@
 using System.Linq;
-using static SpanExtensions.FuzzTests.TestHelper;
+using static SpanExtensions.Tests.Fuzzing.TestHelper;
 
-namespace SpanExtensions.FuzzTests
+namespace SpanExtensions.Tests.Fuzzing
 {
     public static class ReadOnlySpanSplitTests
     {
